+++ openrtb-core/src/test/java/com/google/openrtb/json/Test3Writer.java
@@ -21,10 +21,7 @@ import com.fasterxml.jackson.core.JsonGenerator;
 import java.io.IOException;
 
 /**
- * Sample JSON writer for a regular or repeated extension of scalar type.
- * <p>
- * Nothing different is necessary for repeated extensions; the {@link #write} method
- * will simply be invoked multiple times, for each item in the sequence.
+ * Regular extension: {@code "test3": 99}, scalar type {@code Integer}.
  */
 class Test3Writer extends OpenRtbJsonExtWriter<Integer> {